move migration 20 and 21 to 24 and 25
- migration are run in order, these have been skipped in prod
- hotfix to rename them and make sure they are running
